new code flow in terms of Python class inheritance: SotoStarshade -> SotoStarshadeSKi -> SotoStarshade_ContThrust. moved conversion methods to the class higher up the chain, the stationkeeping class.
